fixes to workoutform updates.
Updated gitignore with vscode launch.json.

added deleteWorkout from api to workoutform.

added some verification we actually have a set to render in renderSets.

move styles imports around in Sets.js

in WorkoutForm, only updating woContext if we do an addWorkout, not with an updateWorkout.  because addWorkout results in an id getting assigned to the workout, update does not.

removed the call to fetchMyAPI() in the Workouts.js > saveWorkout.

remove unnecessary async/await from WoProvider.